# doc/api.yaml
openapi: '3.0.2'
info:
  title: WASAPhoto API Specifications
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! 
    You can upload your photos directly from your PC, and they will be visible to everyone following you.
  version: "1.0"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token for authentication in WASAPhoto.
      bearerFormat: '^[a-z0-9]{64}$'
  schemas:
    ID:
      title: ID
      description: |-
        Autoincrement integer field used for identify a particular resource (a comment, a post, ecc.)
      type: integer
      example: 795739
    Username:
      title: Username
      description: Username of a user
      type: string
      example: "chri_genna02"
      minLength: 8
      maxLength: 16
      pattern: '^[a-zA-Z0-9_-]{8,16}$'
    User:
      title: User
      description: |-
        A user in WASAPhoto is represented by a unique ID, its unique username, name and birthdate.
      properties:
        uuid:
          $ref: '#/components/schemas/ID'
        username:
          $ref: '#/components/schemas/Username'
        profilePic:
          $ref: "#/components/schemas/Photo"
        name:
          description: Real-life name/surname of the user
          type: string
          pattern: ''
          minLength: 2
          maxLength: 32
          example: Christian Gennarelli 
        birthdate:
          $ref: '#/components/schemas/Date'
    UserList:
      title: UserList
      description: |-
        Collection of users. 
        A maximum number of 16 users is returned, according to some backend principle (example: datetime)
      type: array
      items:
        $ref: '#/components/schemas/User'  
      minItems: 0
      maxItems: 16 
    Photo:
      title: Photo
      format: binary
      type: string
      description: |-
        Binary representation of a photo, wrapped in a base64-encoded string. 
        Only multiple profile pics are sent via base64 for efficiency reasons.
        URL pointing to post images will be first sent, and then the client will have the ability
        to obtain the photo via a separate endpoint, i.e. GetPhotoFromURL.
      minLength: 0
      maxLength: 2000000 # 2MB photo
      example: iVBORw0KGgbW8gY2hyaXN0aWFuCghyaXN0aWFuCgw0KGgbW8gYyaXN0aWFuCgw0KGgbW8gW8gYyaXN0aW8gYyaXN0aWF
...
    Date:
      title: date
      description: Date.
      example: "2023-11-29"
      type: string
      format: date
      pattern: '^([0-9]{4})-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$'
      minLength: 10
      maxLength: 10  
    Datetime:
      title: datetime
      description: |-
        Date and time.
      example: "2023-11-29 23:00:00"
      type: string
      format: date-time
      pattern: '^([0-9]{4})-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01]) (0[0-9]|1[0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9])$'
      minLength: 19
      maxLength: 19  
    Error:
      title: Error
      description: |-
        Schema defining what is an error when returned in response of a HTTP requests.
      properties:
        code: # Code returned by the server
          type: string
          description: Status code returned directly by the server(s).
          example: "404"
          minLength: 3
          maxLength: 3
        description:  
          type: string
          description: |- 
            Description of the error in human-readable language.
          example: 'Username not found. Please check that the username is correct, and try again.'
          minLength: 1
          maxLength: 128
          pattern: ''
paths:
...
  /users/{username}/followings/:
    parameters:
      - in: path
        name: username
        schema:
          $ref: '#/components/schemas/Username'
        required: true

    get:
      operationId: getFollowingList
      tags: ['FOLLOW']
      summary: Get the following list
      description: |-
        Return the list of the users that username follows.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: datetime
          schema:
            $ref: '#/components/schemas/Datetime'
          required: false
      responses:
        '200': # OK Found
          description: |-
            The following list if correctly returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
        '204': # No content found
          description: This user does not follow any user.
        '400': # Bad request
          description: Bad request provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': # Unauthenticated
          description: The client is NOT authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403': # Unauthorized
          description: |-
            The authenticated user banned the user provided in the path, or viceversa.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404': # Username not found
          description: Username has not been found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500': # Internal server error.
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
    put: 
      operationId: followUser
      summary: Follow a new user
      tags: ['FOLLOW']
      description: |-
        Add the user provided in the query to the followings list of the authenticated user.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: followed_username
          schema: 
            $ref: '#/components/schemas/Username'
      responses:
        '204': # OK Created
          description: User 'username' successfully started to follow 'following_username'.
        '400': # Bad request
          description: Bad request provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': # Unauthenticated
          description: The client is NOT authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403': # Unauthorized
          description: |-
            The authenticated user cannot follow another user on behalf of some other user.
            That is, the authenticated user's username and the one provided in the path do NOT coincide.
            Alternatively, it could happen that the authenticated user banned the user is trying to follow, or viceversa.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404': # Content not found
          description: The username has not been found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500': # Internal server error.
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{username}/followings/{followed_username}:
    parameters:
        - in: path
          name: username
          schema:
            $ref: '#/components/schemas/Username'
          required: true
        - in: path
          name: followed_username
          schema:
            $ref: '#/components/schemas/Username'
          required: true
    
    delete: 
      operationId: unfollowUser
      tags: ['FOLLOW']
      summary: Unfollow an user
      description: |-
        Delete the authenticated user from the list of followings of the one provided in the path.
      security:
        - BearerAuth: []
      responses:
        '204': # OK
          description: The authenticated user successfully unfollowed the user.
        '400': # Bad request
          description: Bad request provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': # Unauthenticated
          description: The client is NOT authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403': # Unauthorized
          description: |-
            The authenticated user cannot unfollow an user on behalf of another user.
            That is, the authenticated username and the one provided in the path do NOT coincide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500': # Internal server error.
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{username}/followers/:
    parameters:
      - in: path
        name: username
        schema:
          $ref: '#/components/schemas/Username'
        required: true
    
    get: 
      operationId: getFollowersList
      tags: ['FOLLOW']
      summary: Get the list of followers
      description: |-
        Return the list of followers of the given username
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: datetime
          schema:
            $ref: '#/components/schemas/Datetime'
          required: false
      responses:
        '200': # OK
          description: Followers list successfully returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
        '400': # Bad request
          description: Bad request provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': # Unauthenticated
          description: The client is NOT authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403': # Unauthorized
          description: |-
            The authenticated user banned the username provided in the path, or viceversa.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404': # Username or post not found
          description: Either the post to be uncommented or the owner of the post has not been found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500': # Internal server error.
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...